Add files via upload
The issue is, i think the max flow is pointless because the resources available dont change throughout. In example, i used ventillators becuase thats normally limiting. However, it does not matter how many ventillators we have available for examination, discharge, level 2 patients, etc. So, the only place where that number would change is the ventilator arc. Additionally, you have code to reorder so its not limiting. I thought the point was that some were scarce and we cant assume we can reorder if there is a shortage? I do not think longest path would help us so this may be the best option we, as its a complex problem and had to be simplified to dealing with one set of numbers, and use the flow where numbers of supplies not change much throughout.